python_modules/vip_sync/requirements.txt
pyroute2==0.6.4
pyroute2.core==0.6.4
pyroute2.ethtool==0.6.4
pyroute2.ipdb==0.6.4
pyroute2.ipset==0.6.4
pyroute2.ndb==0.6.4
pyroute2.nftables==0.6.4
pyroute2.nslink==0.6.4
PyYAML==5.4.1
